requirements.txt
black==22.3.0
blinker==1.4
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
Flask==2.0.3
Flask-Mail==0.9.1
gunicorn==20.1.0
heroku==0.1.4
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
mypy-extensions==0.4.3
mysql-connector-python==8.0.28
pathspec==0.9.0
Pillow==9.1.0
platformdirs==2.5.1
protobuf==3.19.4
python-dateutil==1.5
python-dotenv==0.20.0
requests==2.28.0
tk==0.1.0
tomli==2.0.1
urllib3==1.26.9
Werkzeug==2.0.3
